% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell.data}
\alias{cell.data}
\title{Time series from 20 HeLa cells}
\format{A data frame with 600 rows and 5 variables:
\describe{
  \item{cell}{Cell identifier}
  \item{timepoint}{Time point data collected}
  \item{x}{x location of the cell (relative to starting location)}
  \item{y}{y location of the cell (relative to starting location)}
  \item{nuclear.roundness}{roundness of the cell's nucleus}
}}
\usage{
cell.data
}
\description{
A dataset containing the X and Y positions and nuclear roundness
of 20 HeLa cells across 30 time points.
}
\keyword{datasets}

