% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{median_abs_diff}
\alias{median_abs_diff}
\title{\code{median_abs_diff} returns the median absolute difference between adjacent data values}
\usage{
median_abs_diff(values, index)
}
\arguments{
\item{values}{A numeric vector of time ordered data values grouped by \code{index}}

\item{index}{A vector of grouping indexes defining distinct time series instances}
}
\value{
A numeric value
}
\description{
\code{median_abs_diff} returns the median absolute difference between adjacent data values
}
\examples{
  num.vecs <- 10
  vec.lengths <- sample(10:15, 10, replace=TRUE)
  index <- rep(1:num.vecs, vec.lengths)
  values <- rep(100*rnorm(num.vecs), vec.lengths) + rnorm(sum(vec.lengths))
  median_abs_diff(values, index)

}

