% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{median_abs_diff_rescale}
\alias{median_abs_diff_rescale}
\title{Return a data frame which contains time series from an input data frame rescaled
by each time series's median absolute temporal fluctuations}
\usage{
median_abs_diff_rescale(data, group.col.name, time.col.name, data.col.names)
}
\arguments{
\item{data}{A data frame containing a time series grouping column, a time
column (assumed uniformly spaced), and multiple time series columns}

\item{group.col.name}{Name of the column in \code{data} which contains the
time series grouping}

\item{time.col.name}{Name of the column in \code{data} which contains the
time for each row}

\item{data.col.names}{Character vector of column names containing simultaneous
time series values}
}
\value{
A data frame containing the same column names specified in the function
  call, but with the data column values rescaled
}
\description{
Return a data frame which contains time series from an input data frame rescaled
by each time series's median absolute temporal fluctuations
}
\examples{
  num.time.series <- 10
  time.series.length <- 25
  num.data.points <-  num.time.series * time.series.length
  data <- data.frame(group=rep(1:num.time.series, time.series.length),
                     time=rep(1:time.series.length, num.time.series),
                     ts1=rnorm(num.data.points),
                     ts2=10*rnorm(num.data.points))
  rescaled.data <- median_abs_diff_rescale(data, "group", "time", c("ts1", "ts2"))

}

